% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleandata.R
\docType{data}
\name{data_file}
\alias{data_file}
\title{Clean Data}
\format{
An object of class \code{character} of length 1.
}
\usage{
data_file
}
\arguments{
\item{data}{}
}
\value{
new dataframe that is cleaned
}
\description{
Clean Data
}
\examples{
\dontrun{
CleanedHousesAvg <- cleandata(HousesAvg)
}

}
\keyword{datasets}
