% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareCityPlots.R
\name{CompareCityPlots}
\alias{CompareCityPlots}
\title{Compare city averages}
\usage{
CompareCityPlots(data, a, b, c)
}
\arguments{
\item{c}{}
}
\value{
3 plots side by side comparing 3 states of your choice
}
\description{
Compare city averages
}
\examples{
\dontrun{
CompareCityPlots(cityonly, "Houston, TX", "Austin, TX", "San Antonio, TX")
}
}
