% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareStatePlots.R
\name{CompareStatePlots}
\alias{CompareStatePlots}
\title{Compare state plots}
\usage{
CompareStatePlots(data, a, b, c)
}
\arguments{
\item{c}{}
}
\value{
3 plots side by side comparing 3 states of your choice
}
\description{
Compare state plots
}
\examples{
\dontrun{
predictStateavg(Stateavg, "TX", 2025)
}
}
