% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonthsCombined.R
\name{MonthsCombined}
\alias{MonthsCombined}
\title{Months Combined}
\usage{
MonthsCombined(data)
}
\arguments{
\item{data}{}
}
\value{
dataset of the months combined
}
\description{
Months Combined
}
\examples{
\dontrun{
YearsAvg <- MonthsCombined(CleanedHousesAvg)
}
}
