% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictCityavg.R
\name{predictCityavg}
\alias{predictCityavg}
\title{Preducted future city value}
\usage{
predictCityavg(data, city, year)
}
\arguments{
\item{year}{}
}
\value{
predicted value of city average
}
\description{
Preducted future city value
}
\examples{
\dontrun{
predictCityavg(Cityavg, "San Antonio, TX", 2025)
}
}
