% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictStateavg.R
\name{predictStateavg}
\alias{predictStateavg}
\title{Predict future state averages}
\usage{
predictStateavg(data, state, year)
}
\arguments{
\item{year}{}
}
\value{
predicted value of state average
}
\description{
Predict future state averages
}
\examples{
\dontrun{
predictStateavg(Stateavg, "TX", 2025)
}
}
